+++ CHANGES.md
@@ -2,14 +2,15 @@ NOTE: as of JNA 4.0, JNA is now dual-licensed under LGPL and AL 2.0 (see LICENSE
 
 NOTE: JNI native support is typically incompatible between minor versions, and almost always incompatible between major versions.
 
-Next Release (5.5.1)
+Next Release (5.6.0)
 ====================
 
 Features
 --------
+* [#1160](https://github.com/java-native-access/jna/issues/1160): Make FileUtils#moveToTrash a varargs method - [@matthiasblaesing](https://github.com/matthiasblaesing).
 
-Features
---------
+Bug Fixes
+---------
 
 
 Release 5.5.0